pyproject.toml
[tool.poetry]
name = "act-safe"
version = "0.1.0"
description = ""
authors = ["Anon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
equinox = "^0.11.1"
jaxlib = { url = "https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.4.23+cuda11.cudnn86-cp310-cp310-manylinux2014_x86_64.whl", extras = [
    "cuda11_pip",
] }
optax = "^0.1.7"
hydra-core = "^1.3.2"
jax = "^0.4.23"
hydra-submitit-launcher = "^1.2.0"
tqdm = "^4.66.2"
tabulate = "^0.9.0"
tensorboardx = "^2.6.2.2"
gymnasium = "^0.29.1"
dm-control = "^1.0.16"
dm-env = "^1.6"
distrax = "^0.1.5"
pillow = "^10.2.0"
moviepy = "^1.0.3"
safe-adaptation-gym = { git = "https://git@github.com/lasgroup/safe-adaptation-gym" }
jmp = { git = "https://github.com/deepmind/jmp" }
tensorboard = "^2.16.2"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
pre-commit = "^3.5.0"
ruff = "^0.1.5"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
types-tabulate = "^0.9.0.20240106"
seaborn = "^0.13.2"
ipykernel = "^6.29.4"


[tool.poetry.group.wandb.dependencies]
wandb = "^0.16.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
ignore = ["E731"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "actsafe/benchmark_suites/humanoid_bench",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
ignore_missing_imports = true
show_column_numbers = true
disallow_untyped_defs = false
disallow_subclassing_any = false
disallow_untyped_calls = false
strict_optional = true
exclude = "actsafe/benchmark_suites/humanoid_bench"

[[tool.mypy.overrides]]
module = ["actsafe.benchmark_suites.humanoid_bench.*"]
ignore_errors = true